+++ src/commands/android-emulator/allowed-values.ts
@@ -1,0 +1,25 @@
+export const deviceIds = {
+  AndroidWearRound: 2,
+  AndroidWearSquare: 4,
+  galaxy_nexus: 5,
+  nexus_4: 7,
+  nexus_5: 8,
+  nexus_5x: 9,
+  nexus_6: 10,
+  nexus_6p: 11,
+  nexus_7: 13,
+  nexus_7_2013: 12,
+  nexus_9: 14,
+  nexus_10: 6,
+  nexus_one: 15,
+  nexus_s: 16,
+  pixel: 17,
+  pixel_c: 18,
+  pixel_xl: 19,
+  tv_720p: 1,
+  tv_1080p: 0,
+};
+
+export const targets = ['google_apis', 'google_apis_playstore'];
+
+export const abis = ['x86', 'armeabi-v7a'];