+++ src/commands/android-emulator/android-emulator.ts
@@ -1,8 +1,9 @@
 import chalk from 'chalk';
-import { createCommand } from '../../command';
-import { assertProperty, replaceVariables } from '../../utils/args';
+import { CommandBuilder } from '../../command';
+import { replaceVariables } from '../../utils/args';
 import { ButlerError } from '../../utils/butler-error';
 import { exists } from '../../utils/file';
+import { abis, deviceIds, targets } from './allowed-values';
 import { createAvd } from './create-avd';
 import { startAvd } from './start-avd';
 
@@ -14,24 +15,56 @@ const ANDROID_HOME_DEFAULTS = [
   'C:\\Users\\{USERNAME}\\AppData\\Local\\Android\\android-sdk',
 ];
 
-export const androidEmulator = createCommand(
-  TAG,
-  {
+interface Config {
+  defaultDevice: string;
+  defaultSdk: string;
+  defaultAbi: string;
+  defaultTarget: string;
+  timeout: number;
+}
+
+interface Args {
+  device: string;
+  sdk: string;
+  abi: string;
+  target: string;
+}
+
+export const androidEmulator = new CommandBuilder<Config, Args>()
+  .name(TAG)
+  .defaultConfig({
     defaultDevice: 'nexus_6p',
     defaultSdk: '27',
     defaultAbi: 'x86',
     defaultTarget: 'google_apis',
     timeout: 30000,
-  },
-  async ({ config, args }) => {
+  })
+  .envVars({
+    ANDROID_HOME: {
+      type: 'string',
+      description: 'Location of Android SDK installation',
+      required: false,
+    },
+  })
+  .config({
+    defaultDevice: { type: Object.keys(deviceIds), description: 'Physical device to emulate.' },
+    defaultSdk: { type: 'string', description: 'Android SDK level.' },
+    defaultAbi: { type: abis, description: 'Architecture to emulate. x86 is obviously faster.' },
+    defaultTarget: { type: targets, description: 'Target extensions to be used. ' +
+        'It\'s recommended to use Google APIs if the application requires Firebase or Play Services.' },
+    timeout: { type: 'number', description: 'Device launch timeout in milliseconds.' },
+  })
+  .args({
+    device: { type: Object.keys(deviceIds), description: 'Physical device to emulate.' },
+    sdk: { type: 'string', description: 'Android SDK level.' },
+    abi: { type: abis, description: 'Architecture to emulate. x86 is obviously faster.' },
+    target: { type: targets, description: 'Target extensions to be used. ' +
+        'It\'s recommended to use Google APIs if the application requires Firebase or Play Services.' },
+  })
+  .execute(async ({ config, args }) => {
     const { ANDROID_HOME } = process.env;
     const { defaultDevice, defaultSdk, defaultAbi, defaultTarget, timeout } = config;
     const { device, sdk, abi, target } = args;
-
-    assertProperty(config, 'defaultDevice', 'string');
-    assertProperty(config, 'defaultSdk', 'string');
-    assertProperty(config, 'defaultAbi', 'string');
-    assertProperty(config, 'defaultTarget', 'string');
 
     let androidHome = ANDROID_HOME;
     if (!androidHome) {
@@ -67,5 +100,5 @@ export const androidEmulator = createCommand(
       timeout,
     });
     console.log(`${TAG}: avd started successfully`);
-  }
-);
+  })
+  .build();