+++ src/commands/android-emulator/create-avd.ts
@@ -3,30 +3,9 @@ import { replaceVariables } from '../../utils/args';
 import { ButlerError } from '../../utils/butler-error';
 import { execAndLog } from '../../utils/execute';
 import { copy, exists, readDir, readFile, writeFile } from '../../utils/file';
+import { deviceIds } from './allowed-values';
 
 const TAG = 'create-avd';
-
-const deviceIds = {
-  AndroidWearRound: 2,
-  AndroidWearSquare: 4,
-  galaxy_nexus: 5,
-  nexus_4: 7,
-  nexus_5: 8,
-  nexus_5x: 9,
-  nexus_6: 10,
-  nexus_6p: 11,
-  nexus_7: 13,
-  nexus_7_2013: 12,
-  nexus_9: 14,
-  nexus_10: 6,
-  nexus_one: 15,
-  nexus_s: 16,
-  pixel: 17,
-  pixel_c: 18,
-  pixel_xl: 19,
-  tv_720p: 1,
-  tv_1080p: 0,
-};
 
 const configurationTemplate = `
 showDeviceFrame=yes